fix: multiple code quality improvements
Fixed critical bugs across multiple modules:

1. Fixed operator precedence bug in sonic.c
   - Corrected type casting order in distance calculation
   - Added explicit parentheses for clarity
   - Affected functions: sonic_measure_mode1(), sonic_measure_mode2(), sonic_measure_mode3()

2. Fixed copy-paste error in uart.c
   - Changed memset(rx_buf, ...) to memset(tx_buf, ...) on line 8
   - Now correctly initializes both rx_buf and tx_buf buffers

3. Fixed display logic bug in rewrite_test/main.c
   - Corrected digit extraction: sonic.distance % 10 / 10 -> sonic.distance / 10
   - Fixed display order: now correctly shows tens digit first

All fixes improve code correctness and prevent potential runtime errors.

+++ 模块/1.4 sonic/main/sonic.c
@@ -32,7 +32,7 @@ uint8_t sonic_measure_mode1() {
     if (TF1 == 1) {
         distance = 255;
     } else {
-        distance = (uint8_t) ((TH1 << 8) | TL1) * 0.017;
+        distance = (uint8_t)(((TH1 << 8) | TL1) * 0.017);
     }
 
     return distance;
@@ -61,7 +61,7 @@ uint8_t sonic_measure_mode2() {
     if (TF1 == 1) {
         distance = 255;
     } else {
-        distance = (uint8_t) ((TH1 << 8) | TL1) * 0.017;
+        distance = (uint8_t)(((TH1 << 8) | TL1) * 0.017);
     }
 
     return distance;
@@ -88,6 +88,6 @@ uint8_t sonic_measure_mode3() {
         return 0;
     } else {
         time_Dis = (CH << 8) | CL;
-        return (time_Dis * 0.0172)-2;
+        return (uint8_t)((time_Dis * 0.0172) - 2);
     }
 }
